requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
sqlalchemy==2.0.37
psycopg2-binary==2.9.10
pydantic-settings==2.7.1
email-validator==2.2.0
PyJWT==2.10.1
python-multipart==0.0.20
pandas==2.2.3
numpy==2.2.2
scikit-learn==1.6.1
statsmodels==0.14.4
openpyxl==3.1.5
openai==2.21.0
